requirements.txt
django==5.0.6
djangorestframework==3.15.2
drf-yasg==1.21.7
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
redis==5.0.7
celery==5.4.0
django-redis==5.4.0
requests==2.32.3
memory_profiler==0.61.0
psycopg2-binary
custom_json_serializer==0.2.7